feat: Add function tokenize to produce an iterator of tokens
Filled the implementation a function to tokenize input into a stream of
tokens which spares the users from having to create a lexer and
advance it manually. This change required fixing an issue with lifetimes
where previously tokens had lifeteime of the mutable borrow of the
Lexer. Now tokens returned from the Lexer have a lifetime equal to that
of the input